Add new level and update enemy spawn settings
Introduced a new level to the main menu scene with updated enemy configurations and abilities. Changed enemySpawnRate type from int to float in Level.cs to allow for more precise spawn rate control. Adjusted existing levels' enemy counts, max enemies on screen, and ability images.

+++ Scripts/Level/Level.cs
@@ -7,5 +7,5 @@ public class Level : MonoBehaviour
     [SerializeField] public Enemy[] enemies;
     [SerializeField] public int[] enemyCountss;
     [SerializeField] public int maxEnemysOnScreen;
-    [SerializeField] public int enemySpawnRate;
+    [SerializeField] public float enemySpawnRate;
 }
